// fe/src/types/generated/UpdateRoleRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PermissionResponse } from './PermissionResponse'

/**
 * Request to update an existing role
 */
export type UpdateRoleRequest = {
    /**
     * Name of the role
     */
    name: string
    /**
     * Optional description
     */
    description?: string
    /**
     * Whether this role grants super admin privileges
     */
    super_admin?: boolean
    /**
     * Direct permissions for this role
     */
    permissions: Array<PermissionResponse>
}